// Optimum Solutions/ClientApp/src/Shared/Models/UpdataData.ts
export interface UpdateData {
    id: string | null;
    salutation: string | null;
    initials: string | null;
    firstname: string | null;
    firstname_ascii: string | null;
    gender: string | null;
    firstname_country_rank: string | null;
    firstname_country_frequency: string | null;
    lastname: string | null;
    lastname_ascii: string | null;
    lastname_country_rank: string | null;
    lastname_country_frequency: string | null;
    email: string | null;
    password: string | null;
    country_code: string | null;
    country_code_alpha: string | null;
    country_name: string | null;
    primary_language_code: string | null;
    primary_language: string | null;
    balance: number | null;
    phone_Number: string | null;
    currency: string | null;
  }
  